pyproject.toml
[project]
name = "ai-daily-paper"
version = "0.1.0"
description = "AI RSS daily report generation pipeline"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.32.0",
  "feedparser>=6.0.11",
  "PyYAML>=6.0.1",
  "markdown>=3.6",
  "beautifulsoup4>=4.12.3",
  "APScheduler>=3.10.4",
  "Pillow>=10.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
